qest_backend_assignment/requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
pymongo
pydantic[email]
